apps/crawler/requirements.txt
selenium==4.15.2
psycopg2-binary==2.9.9
pandas==2.1.3
sqlalchemy==2.0.23
webdriver-manager==4.0.1

